==> R-food-access-lh/1_visualizations.Rmd <==
```{r setup=TRUE}


```

# Map basic input data

```{r}
source("./0_Libraries.R")
# get county boundary and water
la_county <- get_county_boundary() %>%
  st_transform(proj_crs)

la_city <- get_city_boundary(proj_crs) %>%
  st_transform(proj_crs)

# get water
la_water <- area_water(state="CA", county="Los Angeles")%>%
  st_transform(proj_crs) %>%
  clipintersect_boundary(la_city)
  
la_food <- foodpoi_plot %>%
  filter(!is.na(LONGITUDE)) %>%
  st_as_sf(coords=c("LONGITUDE", "LATITUDE"), crs=4326) %>%
  st_transform(proj_crs) #%>%
  #clipintersect_boundary(la_city)

la_ctcent <- get_lac_centroids() %>%
  st_transform(proj_crs) %>% # OSM data is in 4326
  mutate(id=row_number()) 

la_ct_wtcent <- get_lac_weight_centroids() %>%
  st_transform(proj_crs) 



osm_parks <- getbb("Los Angeles County", 
                   display_name_contains = "United States") %>%
            opq() %>% 
            add_osm_feature(key = "leisure", 
                  value=c("park", "nature_reserve", "golf_course")) %>% 
            osmdata_sf()

la_parks <- osm_parks$osm_polygons %>%
  st_transform(proj_crs) %>%
  clipintersect_boundary(la_city)

# get major highways

```

## Mapping study area

```{r}

tm_shape(la_city) +
  tm_polygons(col = "lightgrey", fill_alpha=.2) +
  tm_shape(la_county) +
  tm_borders() + 
  tm_shape(la_water) +
  tm_fill(col="lightblue", lwd=0) +  
  tm_shape(la_parks) +
  tm_polygons(col="lightgreen") + 
  tm_layout(inner.margins = c(0.1, .1, 0.2, 0.1)) +
  tm_scale_bar()

```

## Mapping food access poi

```{r}

tm_shape(la_city) +
  tm_borders(col="darkblue") +
  tm_shape(la_food %>% filter(category!="SPF" | category!="Not included"), alpha=.9) +
  tm_dots(col="category", size = 0.02) +  
  tm_shape(la_water) +
  tm_fill(col="lightblue", lwd=0) + 
  tm_scale_bar(position=c("LEFT", "BOTTOM")) 

```

## Mapping population weighted census tract centroids

```{r}
# remotes::install_version("tmap", version = "3.3-4") fix for tmap view
# detach("package:tmap", unload=TRUE)

tmap_mode("plot")

boundary1 <- st_buffer(la_city, -200) 

la_city_ct <- la_ct %>% filter((lengths(st_intersects(., boundary1)) > 0))

tm_shape(la_city_ct) +
  tm_polygons(col="white") +
  tm_shape(la_ctcent %>% filter((lengths(st_intersects(., la_city)) > 0))) +
  tm_dots(col = "red", size = 0.02) +
  tm_shape(la_ct_wtcent %>% filter((lengths(st_intersects(., la_city)) > 0))) +
  tm_dots(col = "purple", size = 0.02) +  
  tm_shape(la_water) +
  tm_fill(col="lightblue", lwd=0) + 
  tm_layout(legend.position = c("left", "bottom")) 

```

# Inspecting census tract point datasets for issues

We inspected the census tracts with strange behavior ID = 980016, 990200, 599100 (in the middle of ocean) using census reporter because there were issues plotting these points

980016 - split into two (do not erase water - introduces invalid geoms, handled in code) ✔️

9902 - zero population (remove all cts with 0 aland) ✔️

5991 - catalina island (remove)

```{tract <- la_ct[la_ct$TRACTCE==980016,]}

print(tract$TRACTCE[[1]])  # Find adjacent tracts  

dot <- la_ctcent_dat[la_ctcent_dat$TRACTCE==980016,] 
dotwt <- la_ct_wtcent_dat[la_ct_wtcent_dat$TRACTCE==980016,]

# Plot the selected and adjacent tracts 
figure <- tm_shape(dot) + tm_dots(col = "red", size = 0.01) + 
  tm_shape(dotwt) + 
  tm_dots(col = "green", size = 0.01) + 
  tm_shape(la_ct) + 
  tm_borders() + 
  tm_layout(legend.position = c("left", "bottom"))

print(figure)

```

980030 - no weighted CT point due to zero population (remove all with 0 population) ✔️

we find all empty points in la_ct_wtcent_dat and inspect their corresponding blocks to find they are all population 0; adjusted to remove 0 pop CTs ✔️

```{r}
empty <- la_ct_wtcent_dat[st_is_empty(la_ct_wtcent_dat),]

empty_blocks <- la_cb[la_cb$TRACTCE %in% empty$TRACTCE,]
```

# Mapping POI and population-weighted centroids data

```{r}
tm_shape(la_county) +
  tm_borders() +
  tm_shape(food_23_24_SSI) +
  tm_dots(col = "red", size = 0.01) +
  tm_shape(la_ct) +
  tm_borders() +
  tm_shape(la_ctcent_dat) +
  tm_dots(col = "purple", size = 0.01) +
  tm_layout(legend.position = c("left", "bottom"))

```

## Map terrain elevation and street network features

```{r}

elev <- get_dem("socal", proj_crs)

tmap_options(check.and.fix = TRUE)

tm_shape(elev) +
  tm_raster(palette = terrain.colors(100), style = "cont", title = "Elevation (m)")+ 
  tm_shape(la_county) + 
   tm_borders(col = "black", alpha=.5)


```

## Visualizing OSM

```{r}
  osm_lines <- get_osm(bbox=lac_bbox)
  ht = c("raceway") # highway types to remove
  # highway types https://wiki.openstreetmap.org/wiki/Key:highway#Roads
  osm_paths = osm_lines[!(osm_lines$highway %in% ht), ]
  plot(osm_paths["highway"], key.pos = 1)
  
  rm(osm_lines, osm_paths)
  
  #not enough heap space to run this 
  #network <- street_network_to_sf(r5r_core)


```
